requirements.txt
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
Fiona==1.8.4
geopandas==0.4.1
idna==2.8
munch==2.3.2
numpy==1.16.2
opencv-python==4.0.0.21
pandas==0.24.1
Pillow==5.4.1
pyproj==2.0.2
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
Shapely==1.6.4.post2
six==1.12.0
urllib3==1.24.1
